refactor(fw): move TransactionType enum to ethereum_test_types module (#1673)
* refactor(fw): move TransactionType enum to ethereum_test_types module

* fix: tox

---------
