+++ whitelist.txt
@@ -389,6 +389,7 @@ returncontract
 returndata
 returndatacopy
 returndatasize
+reusability
 rlp
 rootdir
 rpc
